requirements.txt
streamlit
fastapi
tensorflow
pillow
numpy
uvicorn
python-multipart